beets-13 re-organized checkin components
